+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Unreleased
 
+  * Fix ICE disconnections with Chrome v144+ when not using ICE Lite #846
   * Downgrade noisy BWE logs #847
   * Fix H265 PACI depacketizer panic on short payload #849